Use u64 for OffsetType::MAX in truncate_buffer

+++ crates/polars-compute/src/cast/utf8_to.rs
@@ -77,7 +77,7 @@ type OffsetType = i8;
 fn truncate_buffer(buf: &Buffer<u8>) -> Buffer<u8> {
     // * 2, as it must be able to hold u32::MAX offset + u32::MAX len.
     buf.clone()
-        .sliced(0, std::cmp::min(buf.len(), OffsetType::MAX as usize * 2))
+        .sliced(0, std::cmp::min(buf.len(), ((OffsetType::MAX as u64) * 2) as usize))
 }
 
 pub fn binary_to_binview<O: Offset>(arr: &BinaryArray<O>) -> BinaryViewArray {
